+++ sample/frontend/src/containers/Entrance/index.tsx
@@ -9,7 +9,9 @@ import LoginEntranceFormFieldGroup from '@/components/LoginEntranceFormFieldGrou
 const Entrance: React.FC<Record<string, never>> = () => {
   const params: { roomId: string } = useParams();
   // eslint-disable-next-line @typescript-eslint/naming-convention
-  const { video_bitrate, share_bitrate, default_layout, enable_video, enable_audio, use_dummy_device, bitrate_reservation_mbps, room_type, video_codec, is_debug } = qs.parse(window.location.search);
+  const { video_bitrate, share_bitrate, default_layout, enable_video, enable_audio, audio_mute_type, use_dummy_device, bitrate_reservation_mbps, room_type, video_codec, is_debug } = qs.parse(
+    window.location.search
+  );
   const [roomId, setRoomId] = useState<string>('');
   const [username, setUsername] = useState<string>('');
   const onSubmitSuccess = (): void => {
@@ -30,6 +32,9 @@ const Entrance: React.FC<Record<string, never>> = () => {
     }
     if (enable_audio) {
       uriPath += `&enable_audio=${enable_audio}`;
+    }
+    if (audio_mute_type) {
+      uriPath += `&audio_mute_type=${audio_mute_type}`;
     }
     if (use_dummy_device) {
       uriPath += `&use_dummy_device=${use_dummy_device}`;